WIP refactoring recently_sent_txs. Compilation fails.
